+++ src/deskTask/Desk.java
@@ -1,4 +1,4 @@
-package thirdLesson;
+package deskTask;
 
 import java.util.ArrayList;
 import java.util.List;
