+++ src/deskTask/MainCards.java
@@ -1,4 +1,5 @@
-package thirdLesson;
+package deskTask;
+
 
 
 public class MainCards {
